pyproject.toml
[tool.poetry]
name = "codementor-article-rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
rag-chroma = {path = "packages/rag-chroma", develop = true}
langchain-openai = "^0.0.2.post1"
langchain-community = "^0.0.12"
bs4 = "^0.0.1"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
